Profile now displays all user info who is logged in.
