+++ src/contexts/ProfileContext.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const ProfileContext = createContext();
+
+export default ProfileContext;
